OP-1154 refactored Long method using Extract Class
